[REF] refactor code in order to have more generic code in search engine
