+++ connector_search_engine/unit/__init__.py
@@ -3,5 +3,6 @@
 # Sébastien BEAU <sebastien.beau@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import models
-from . import unit
+
+from . import exporter
+from . import adapter
