+++ solerp/unit/adapter.py
@@ -19,26 +19,23 @@
 #
 ##############################################################################
 
-from openerp.addons.connector.unit.backend_adapter import BackendAdapter
+from openerp.addons.connector_search_engine.unit.adapter import\
+    SearchEngineAdapter
 from sunburnt import SolrInterface
 from ..backend import solr
 
 
 @solr
-class SolrAdapter(BackendAdapter):
+class SolrAdapter(SearchEngineAdapter):
     _model_name = None
 
     __solr_pool = {}  # pool of connection for solr
 
-    @classmethod
-    def match(cls, session, model):
-        return True  # We are a generic exporter; how cool is that?
-
     def __init__(self, connector_env):
-        self.location = connector_env.backend_record.location
-        if not self.__solr_pool.get(self.location):
-            self.__solr_pool[self.location] = SolrInterface(self.location)
-        self.conn = self.__solr_pool[self.location]
+        location = connector_env.backend_record.location
+        if not self.__solr_pool.get(location):
+            self.__solr_pool[location] = SolrInterface(location)
+        self.conn = self.__solr_pool[location]
 
     def add(self, datas):
         self.conn.add(datas, len(datas))
